Refactor to fix gwtc
Split S3Request into server side and client side, and remove ZoneDateTime references client side.
Move bat stuff to client side
Add hmac-sha256 to Hasher
