+++ src/peergos/server/storage/S3BucketCopy.java
@@ -1,7 +1,10 @@
 package peergos.server.storage;
 
+import peergos.server.*;
 import peergos.server.util.*;
+import peergos.shared.crypto.hash.*;
 import peergos.shared.storage.*;
+import peergos.shared.storage.auth.*;
 
 import java.io.*;
 import java.time.*;
@@ -15,25 +18,26 @@ public class S3BucketCopy {
 
     private static final Logger LOG = Logger.getGlobal();
 
-    private static void applyToAllInRange(Consumer<S3Request.ObjectMetadata> processor,
+    private static void applyToAllInRange(Consumer<S3AdminRequests.ObjectMetadata> processor,
                                           String startPrefix,
                                           Optional<String> endPrefix,
                                           S3Config config,
-                                          AtomicLong counter) {
+                                          AtomicLong counter,
+                                          Hasher h) {
         try {
             Optional<String> continuationToken = Optional.empty();
-            S3Request.ListObjectsReply result;
+            S3AdminRequests.ListObjectsReply result;
             do {
-                result = S3Request.listObjects(startPrefix, 1_000, continuationToken,
+                result = S3AdminRequests.listObjects(startPrefix, 1_000, continuationToken,
                         ZonedDateTime.now(), config.getHost(), config.region, config.accessKey, config.secretKey, url -> {
                             try {
                                 return HttpUtil.get(url);
                             } catch (IOException e) {
                                 throw new RuntimeException(e);
                             }
-                        });
+                        }, S3AdminRequests.builder::get, h);
 
-                for (S3Request.ObjectMetadata objectSummary : result.objects) {
+                for (S3AdminRequests.ObjectMetadata objectSummary : result.objects) {
                     if (objectSummary.key.endsWith("/")) {
                         LOG.fine(" - " + objectSummary.key + "  " + "(directory)");
                         continue;
@@ -54,17 +58,18 @@ public class S3BucketCopy {
         }
     }
 
-    private static Set<String> getFilenames(S3Config config) {
+    private static Set<String> getFilenames(S3Config config, Hasher h) {
         Set<String> results = new HashSet<>();
-        applyToAllInRange(obj -> results.add(obj.key), "", Optional.empty(), config, new AtomicLong(0));
+        applyToAllInRange(obj -> results.add(obj.key), "", Optional.empty(), config, new AtomicLong(0), h);
         return results;
     }
 
     private static void copyObject(String key,
                                    String sourceBucket,
-                                   S3Config config) {
-        PresignedUrl copyUrl = S3Request.preSignCopy(sourceBucket, key, key,
-                ZonedDateTime.now(), config.getHost(), Collections.emptyMap(), config.region, config.accessKey, config.secretKey);
+                                   S3Config config,
+                                   Hasher h) {
+        PresignedUrl copyUrl = S3Request.preSignCopy(sourceBucket, key, key, S3AdminRequests.asAwsDate(ZonedDateTime.now()), config.getHost(),
+                Collections.emptyMap(), config.region, config.accessKey, config.secretKey, h).join();
         try {
             System.out.println("Copying s3://"+sourceBucket + "/" + key + " to s3://" + config.bucket);
             String res = new String(HttpUtil.put(copyUrl, new byte[0]));
@@ -81,9 +86,10 @@ public class S3BucketCopy {
                                   S3Config destConfig,
                                   AtomicLong counter,
                                   AtomicLong copyCounter,
-                                  int parallelism) {
+                                  int parallelism,
+                                  Hasher h) {
         System.out.println("Listing destination bucket...");
-        Set<String> targetKeys = getFilenames(destConfig);
+        Set<String> targetKeys = getFilenames(destConfig, h);
         ForkJoinPool pool = new ForkJoinPool(parallelism);
         System.out.println("Copying objects...");
         applyToAllInRange(obj -> {
@@ -91,9 +97,9 @@ public class S3BucketCopy {
                 while (pool.getQueuedSubmissionCount() > 100)
                     try {Thread.sleep(100);} catch (InterruptedException e) {}
                 copyCounter.incrementAndGet();
-                pool.submit(() -> copyObject(obj.key, sourceConfig.bucket, destConfig));
+                pool.submit(() -> copyObject(obj.key, sourceConfig.bucket, destConfig, h));
             }
-        }, startPrefix, endPrefix, sourceConfig, counter);
+        }, startPrefix, endPrefix, sourceConfig, counter, h);
         System.out.println("Objects copied: " + copyCounter.get());
     }
 
@@ -108,6 +114,7 @@ public class S3BucketCopy {
         Optional<String> endPrefix = Optional.empty();
 
         System.out.println("Copying S3 bucket " + sourceBucket + " to " + destConfig.bucket);
-        copyRange(startPrefix, endPrefix, sourceConfig, destConfig, new AtomicLong(0), new AtomicLong(0), a.getInt("parallelism"));
+        copyRange(startPrefix, endPrefix, sourceConfig, destConfig, new AtomicLong(0),
+                new AtomicLong(0), a.getInt("parallelism"), Main.initCrypto().hasher);
     }
 }
