+++ src/peergos/shared/storage/DelegatingStorage.java
@@ -1,10 +1,10 @@
 package peergos.shared.storage;
 
-import peergos.server.storage.auth.*;
 import peergos.shared.cbor.*;
 import peergos.shared.crypto.hash.*;
 import peergos.shared.io.ipfs.cid.*;
 import peergos.shared.io.ipfs.multihash.*;
+import peergos.shared.storage.auth.*;
 import peergos.shared.user.fs.*;
 import peergos.shared.util.*;
 
@@ -106,9 +106,10 @@ public abstract class DelegatingStorage implements ContentAddressedStorage {
     @Override
     public CompletableFuture<List<FragmentWithHash>> downloadFragments(List<Cid> hashes,
                                                                        List<BatWithId> bats,
+                                                                       Hasher h,
                                                                        ProgressConsumer<Long> monitor,
                                                                        double spaceIncreaseFactor) {
-        return target.downloadFragments(hashes, bats, monitor, spaceIncreaseFactor);
+        return target.downloadFragments(hashes, bats, h, monitor, spaceIncreaseFactor);
     }
 
     @Override
