+++ src/peergos/shared/storage/auth/BatWithId.java
@@ -1,4 +1,4 @@
-package peergos.server.storage.auth;
+package peergos.shared.storage.auth;
 
 import peergos.shared.cbor.*;
 import peergos.shared.io.ipfs.cid.*;
